fixed MainActivity
