+++ app/src/main/java/com/cmput301f17t11/cupofjava/Views/AllHabitViewActivity.java
@@ -82,82 +82,6 @@ public class AllHabitViewActivity extends Fragment {
         }
 
 
-        //handle bottom navigation bar
-        final BottomNavigationView bottomNavigationView = (BottomNavigationView) view.findViewById(R.id.bottom_navigation_today);
-        BottomNavigationViewHelper.disableShiftMode(bottomNavigationView);
-        Menu menu = bottomNavigationView.getMenu();
-        MenuItem menuItem = menu.getItem(2);
-        menuItem.setChecked(true);
-
-        bottomNavigationView.setOnNavigationItemSelectedListener(new BottomNavigationView.OnNavigationItemSelectedListener() {
-            @Override
-            public boolean onNavigationItemSelected(@NonNull MenuItem item) {
-                switch(item.getItemId()) {
-
-                    case R.id.action_timeline:
-                        Bundle bundle2 = new Bundle();
-                        bundle2.putString("userName", userName);
-                        HomeFragment fragment2 = new HomeFragment();
-                        fragment2.setArguments(bundle2);
-                        FragmentManager fragmentManager2 = getFragmentManager();
-                        FragmentTransaction fragmentTransaction2 = fragmentManager2.beginTransaction();
-                        fragmentTransaction2.replace(R.id.frame, fragment2).addToBackStack(null);
-                        fragmentTransaction2.commit();
-                        //intent2.putExtra("userIndex", userIndex);
-                        break;
-                    case R.id.action_today:
-                        Bundle bundle3 = new Bundle();
-                        bundle3.putString("userName", userName);
-                        TodayViewActivity fragment3 = new TodayViewActivity();
-                        fragment3.setArguments(bundle3);
-                        FragmentManager fragmentManager3 = getFragmentManager();
-                        FragmentTransaction fragmentTransaction3 = fragmentManager3.beginTransaction();
-                        fragmentTransaction3.replace(R.id.frame, fragment3).addToBackStack(null);
-                        fragmentTransaction3.commit();
-                        //intent2.putExtra("userIndex", userIndex);
-                        break;
-                    case R.id.action_all_habits:
-                        break;
-                    case R.id.add_habit_or_habit_event:
-                        AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
-                        builder.setTitle("Add New")
-                                .setNegativeButton("New Habit", new DialogInterface.OnClickListener() {
-                                    @Override
-                                    public void onClick(DialogInterface dialog, int which) {
-                                        Intent intent4 = new Intent(getActivity(), NewHabitActivity.class);
-                                        intent4.putExtra("userName", userName);
-                                        startActivity(intent4);
-                                    }
-                                })
-                                .setPositiveButton("New Event", new DialogInterface.OnClickListener() {
-                                    @Override
-                                    public void onClick(DialogInterface dialog, int which) {
-                                        Intent intent5 = new Intent(getActivity(), NewHabitEventActivity.class);
-                                        intent5.putExtra("userName", userName);
-                                        startActivity(intent5);
-                                    }
-                                });
-
-
-                        AlertDialog dialog = builder.create();
-                        dialog.show();
-                        break;
-
-                    case R.id.action_friends:
-                        Bundle bundle4 = new Bundle();
-                        bundle4.putString("userName", userName);
-                        SocialFragment fragment4 = new SocialFragment();
-                        fragment4.setArguments(bundle4);
-                        FragmentManager fragmentManager4 = getFragmentManager();
-                        FragmentTransaction fragmentTransaction4 = fragmentManager4.beginTransaction();
-                        fragmentTransaction4.replace(R.id.frame, fragment4).addToBackStack(null);
-                        fragmentTransaction4.commit();
-                        //intent2.putExtra("userIndex", userIndex);
-                        break;
-                }
-                return false;
-            }
-        });
 
 
         //set up TextView and ListView
