+++ app/src/main/java/com/cmput301f17t11/cupofjava/Views/TodayViewActivity.java
@@ -100,73 +100,6 @@ public class TodayViewActivity extends Fragment {
         MenuItem menuItem = menu.getItem(1);
         menuItem.setChecked(true);
 
-        bottomNavigationView.setOnNavigationItemSelectedListener(new BottomNavigationView.OnNavigationItemSelectedListener() {
-            @Override
-            public boolean onNavigationItemSelected(@NonNull MenuItem item) {
-                switch(item.getItemId()) {
-
-                    case R.id.action_timeline:
-                        Bundle bundle = new Bundle();
-                        bundle.putString("userName", userName);
-                        HomeFragment fragment = new HomeFragment();
-                        fragment.setArguments(bundle);
-                        FragmentManager fragmentManager = getFragmentManager();
-                        FragmentTransaction fragmentTransaction = fragmentManager.beginTransaction();
-                        fragmentTransaction.replace(R.id.frame, fragment).addToBackStack(null);
-                        fragmentTransaction.commit();
-                        //intent2.putExtra("userIndex", userIndex);
-                        break;
-                    case R.id.action_today:
-                        break;
-                    case R.id.action_all_habits:
-                        Bundle bundle4 = new Bundle();
-                        bundle4.putString("userName", userName);
-                        AllHabitViewActivity fragment4 = new AllHabitViewActivity();
-                        fragment4.setArguments(bundle4);
-                        FragmentManager fragmentManager4 = getFragmentManager();
-                        FragmentTransaction fragmentTransaction4 = fragmentManager4.beginTransaction();
-                        fragmentTransaction4.replace(R.id.frame, fragment4).addToBackStack(null);
-                        fragmentTransaction4.commit();
-                        //intent2.putExtra("userIndex", userIndex);
-                        break;
-                    case R.id.add_habit_or_habit_event:
-                        AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
-                        builder.setTitle("Add New")
-                                .setNegativeButton("New Habit", new DialogInterface.OnClickListener() {
-                                    @Override
-                                    public void onClick(DialogInterface dialog, int which) {
-                                        Intent intent4 = new Intent(getActivity(), NewHabitActivity.class);
-                                        intent4.putExtra("userName", userName);
-                                        startActivity(intent4);
-                                    }
-                                })
-                                .setPositiveButton("New Event", new DialogInterface.OnClickListener() {
-                                    @Override
-                                    public void onClick(DialogInterface dialog, int which) {
-                                        Intent intent5 = new Intent(getActivity(), NewHabitEventActivity.class);
-                                        intent5.putExtra("userName", userName);
-                                        startActivity(intent5);
-                                    }
-                                });
-
-
-                        AlertDialog dialog = builder.create();
-                        dialog.show();
-                        break;
-                    case R.id.action_friends:
-                        Bundle bundle3 = new Bundle();
-                        bundle3.putString("userName", userName);
-                        SocialFragment fragment3 = new SocialFragment();
-                        fragment3.setArguments(bundle3);
-                        FragmentManager fragmentManager3 = getFragmentManager();
-                        FragmentTransaction fragmentTransaction3 = fragmentManager3.beginTransaction();
-                        fragmentTransaction3.replace(R.id.frame, fragment3).addToBackStack(null);
-                        fragmentTransaction3.commit();
-                        //intent2.putExtra("userIndex", userIndex);
-                }
-                return false;
-            }
-        });
 
         //set up the TextView and ListView
         this.textView = (TextView) view.findViewById(R.id.SelfProfileHeadingTextView);
