+++ app/src/main/res/layout/activity_main.xml
@@ -13,7 +13,7 @@
 
 
         <android.support.design.widget.BottomNavigationView
-            android:id="@+id/bottom_navigation_today"
+            android:id="@+id/menu"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_alignParentBottom="true"
